Decompose site content/scripts and fix broken changelog rendering
Fixes a real production bug found while decomposing changelog-page.ts:
changelog.astro's <script define:vars> tag contained a relative
`import`, but Astro's define:vars scripts aren't run through Vite's
import resolution, so it 404s at runtime. The release list has been
silently stuck on "Loading release notes..." with no console error.
Fixed by splitting into two script tags: a define:vars script that
only stashes server-computed data on window, and a plain <script>
(mirroring Header.astro's already-working pattern) that imports
normally and reads the data back off window. Verified fixed live: the
release list and timeline now render real content with zero console
errors.

- Delete the dead site/src/scripts/docs-header.ts (its initDocsHeader
  had drifted from Header.astro's live inline script with unshipped
  features - mobile section-nav repositioning, an anchor-link copy
  enhancer) and replace it with a typed, verbatim extraction of
  Header.astro's actual live script, imported via a plain <script>.
  Verified via a live scroll-event test that data-docs-scrolled still
  updates correctly. The unshipped features are not resurrected here -
  promoting them would be a behavior change beyond decomposition scope.
- Split site/src/data/site.ts (517 lines) into
  site/src/data/site-content/{types,version,meta,hero,surfaces,install,
  faq,footer}.ts, reassembled by a thin site.ts all existing consumers
  still import unchanged. Verified via a side-by-side build comparison
  against the pre-split baseline. Deliberately kept FAQ's `a` and
  `aHtml` both explicitly authored rather than deriving `a` from
  `aHtml`: `a` is embedded verbatim into JSON-LD structured data, and
  some `aHtml` entries contain structural markup that would garble the
  derived plain text.
- Split site/src/scripts/changelog-page.ts (737 lines) into
  changelog/{types,markdown-inline,summarize,markdown-block,timeline,
  release-list}.ts with real types throughout.
- Delete the orphaned scripts/split-site-css.mjs and
  scripts/patch-site-scripts.mjs (patch-site-scripts.mjs is in fact the
  source of the changelog script-tag bug above).
- Wire check-file-size.mjs and test:a11y into site.yml. Before wiring
  in the a11y gate, ran it and fixed 2 real pre-existing violations
  that would have made it immediately red: scrollable-region-focusable
  on JS-created table-scroll wrappers and a static code block (add
  tabIndex/role/aria-label), and a missing accessible name on the
  logo-only nav link (add aria-label). Re-verified 0 violations across
  all 7 routes after fixing.

Fourth in the sequenced audit-remediation batch.
