+++ site/src/components/Nav.astro
@@ -15,7 +15,7 @@ const links = [
 ---
 <nav data-site-nav data-open="false">
   <div class="nav-inner">
-    <a class="mark" href="/">
+    <a class="mark" href="/" aria-label="NetsCLI home">
       <img src={branding.wordmark} alt="" width="160" height="32" />
     </a>
     <button
